requirements.txt
numpy
opencv-python
opencv-contrib-python
pyserial==3.5
